Rename Details subfolder and files
